Add warning re build warning when R3.14.12.5 changed roff from signed to unsigned.  Possible conversion errors.

+++ Bx9000_MBTApp/src/devKL2512.c
@@ -16,7 +16,11 @@ static long lincvt_ao_KL2512(struct aoRecord	*pao, int after)
 	 * Linear conversion slope calculated from eng units full - eng units low
 	 */
 	pao->eslo = (	(pao->eguf - pao->egul)
-				/	(float)(RAW_DUTY_CYCLE_100 - RAW_DUTY_CYCLE_0) );
+				/	(double)(RAW_DUTY_CYCLE_100 - RAW_DUTY_CYCLE_0) );
+	/* This line throws a warning in R3.15.5 re overflow during conversion.
+	 * In R3.14.12.4, roff is epicsInt32, but in R3.14.12.5 it changed to epicsUint32.
+	 * We may need to revisit this conversion!
+	 */
 	pao->roff = -	(float) RAW_DUTY_CYCLE_0;
 	return(0);
 }
